fix(APISIMP-ADMIN-AUDIT-PAGECAP): fix CWE-190 int overflow in listOrphans page*pageSize arithmetic
InstanceAdminRest.java:214 called listOrphans(page * pageSize, pageSize) where both
operands are int — overflow at large page values.  Fix: cast first operand to long
before multiplication.  Propagate long skip down through PermissionAuditService and
update test stubs to use explicit long literals (0L, 100L).

+++ backend/src/main/java/de/dlr/shepard/v2/admin/resources/InstanceAdminRest.java
@@ -211,7 +211,7 @@ public class InstanceAdminRest {
   ) {
     requireInstanceAdmin(securityContext);
     long total = permissionAuditService.countOrphans();
-    List<PermissionAuditEntryIO> orphans = permissionAuditService.listOrphans(page * pageSize, pageSize);
+    List<PermissionAuditEntryIO> orphans = permissionAuditService.listOrphans((long) page * pageSize, pageSize);
     return Response.ok(new PagedResponseIO<>(orphans, total, page, pageSize))
       .header("X-Total-Count", total)
       .build();
